A shot-in-the-dark-I-hope-this-compiles-and-don't-have-to-buy-yum-yums type of commit
git-svn-id: svn+ssh://dash.maidsafe.net/srv/svn/dawn/trunk@3582 6165b5c9-d236-0410-b717-85aa93a55b26

+++ src/tests/maidsafe/testvaultservice.cc
@@ -798,7 +798,7 @@ TEST_F(VaultServicesTest, BEH_MAID_ServicesStoreChunkReference) {
 
     std::vector<std::string> values;
     ASSERT_TRUE(knode_->FindValueLocal(chunkname, &values));
-    ASSERT_EQ(1, values.size());
+    ASSERT_EQ(size_t(1), values.size());
     ASSERT_EQ(ser_signed_value, values[0]);
   }
 }
